Modified view a tad. Doesn't work need to learn cascading drop down lists.

+++ AircraftBalance/AircraftBalance.Models/AircraftViewModel.cs
@@ -9,7 +9,7 @@ namespace AircraftBalance.Models
         public int Plane { get; set; }
         public IEnumerable<Aircraft> Aircraft { get; set; }
         public IEnumerable<PayloadItem> PayloadItems { get; set; }
-
+        
         //Basic Properties
         public int AircraftId { get; set; }
         [Display(Name = "Aircraft Name")]
@@ -18,31 +18,31 @@ namespace AircraftBalance.Models
         public string AircraftMake { get; set; }
         [Display(Name = "Aircraft Model")]
         public string AircraftModel { get; set; }
-        public int BasicEmptyWeight { get; set; }
-        public float BasicEmptyWeightArm { get; set; }
-        public float BasicEmptyweightMoment { get; set; }
-        public int ZeroFuelWeight { get; set; }
-        public float ZeroFuelWeightArm { get; set; }
-        public float ZeroFuelWeightMoment { get; set; }
+        public double BasicEmptyWeight { get; set; }
+        public double BasicEmptyWeightArm { get; set; }
+        public double BasicEmptyweightMoment { get; set; }
+        public double ZeroFuelWeight { get; set; }
+        public double ZeroFuelWeightArm { get; set; }
+        public double ZeroFuelWeightMoment { get; set; }
         [Display(Name = "Fuel Load")]
-        public int FuelLoadWeight { get; set; }
-        public float FuelLoadMoment { get; set; }
-        public float FuelLoadArm { get; set; }
-        public int RampWeight { get; set; }
-        public float RampWeightArm { get; set; }
-        public float RampWeightMoment { get; set; }
-        public int LessFuelForTaxiWeight { get; set; }
+        public double FuelLoadWeight { get; set; }
+        public double FuelLoadMoment { get; set; }
+        public double FuelLoadArm { get; set; }
+        public double RampWeight { get; set; }
+        public double RampWeightArm { get; set; }
+        public double RampWeightMoment { get; set; }
+        public double LessFuelForTaxiWeight { get; set; }
         [Display(Name = "Less Fuel For Taxi")]
-        public float LessFuelForTaxiArm { get; set; }
-        public float LessFuelForTaxiMoment { get; set; }
-        public int TakeOffWeight { get; set; }
-        public float TakeOffWeightArm { get; set; }
-        public float TakeOffWeightMoment { get; set; }
+        public double LessFuelForTaxiArm { get; set; }
+        public double LessFuelForTaxiMoment { get; set; }
+        public double TakeOffWeight { get; set; }
+        public double TakeOffWeightArm { get; set; }
+        public double TakeOffWeightMoment { get; set; }
         [Display(Name = "Less Fuel to Destination")]
-        public int LessFuelToDestination { get; set; }
-        public int LandingWeight { get; set; }
-        public float LandingWeightArm { get; set; }
-        public float LandingWeightMoment { get; set; }
+        public double LessFuelToDestination { get; set; }
+        public double LandingWeight { get; set; }
+        public double LandingWeightArm { get; set; }
+        public double LandingWeightMoment { get; set; }
 
         //Max value properties
         [Display(Name = "Max Landing Weight")]
@@ -54,30 +54,30 @@ namespace AircraftBalance.Models
         [Display(Name = "Max Zero Fuel Weight")]
         public int MaxZeroFuelWeight { get; set; }
         [Display(Name = "Forward Limit")]
-        public float ZeroFuelWeightForwardCG { get; set; }
+        public double ZeroFuelWeightForwardCG { get; set; }
         [Display(Name = "Aft Limit")]
-        public float ZeroFuelWeightAftCG { get; set; }
+        public double ZeroFuelWeightAftCG { get; set; }
         [Display(Name = "Forward Limit")]
-        public float TakeoffWeightForwardCG { get; set; }
+        public double TakeoffWeightForwardCG { get; set; }
         [Display(Name = "Aft Limit")]
-        public float TakeoffWeightAftCG { get; set; }
+        public double TakeoffWeightAftCG { get; set; }
         [Display(Name = "Forward Limit")]
-        public float LandingWeightForwardCG { get; set; }
+        public double LandingWeightForwardCG { get; set; }
         [Display(Name = "Aft Limit")]
-        public float LandingWeightAftCG { get; set; }
+        public double LandingWeightAftCG { get; set; }
 
         //Payload Properties
         public int PayloadId { get; set; }
-        public float PayloadWeight { get; set; }
-        public float PayloadArm { get; set; }
-        public float PayloadMoment { get; set; }
+        public double PayloadWeight { get; set; }
+        public double PayloadArm { get; set; }
+        public double PayloadMoment { get; set; }
 
         //Payload Item Properties
         public int PayloadItemId { get; set; }
         public string PayloadItemName { get; set; }
-        public float PayloadItemWeight { get; set; }
-        public float PayloadItemMoment { get; set; }
-        public float PayloadItemArm { get; set; }
+        public double PayloadItemWeight { get; set; }
+        public double PayloadItemMoment { get; set; }
+        public double PayloadItemArm { get; set; }
 
     }
 }
